+++ src/SharpSvn/SvnStreamWrapper.cpp
@@ -26,9 +26,9 @@ static svn_error_t *svnStreamRead(void *baton, char *buffer, apr_size_t *len)
 
 	SvnStreamWrapper^ sw = AprBaton<SvnStreamWrapper^>::Get((IntPtr)baton);
 
-	array<unsigned char>^ bytes = gcnew array<unsigned char>(*len);
+	array<unsigned char>^ bytes = gcnew array<unsigned char>((int)*len);
 
-	int count = sw->Stream->Read(bytes, 0, *len);
+	int count = sw->Stream->Read(bytes, 0, (int)*len);
 
 	Marshal::Copy(bytes, 0, (IntPtr)buffer, count);
 	*len = count;
@@ -46,7 +46,7 @@ static svn_error_t *svnStreamWrite(void *baton, const char *data, apr_size_t *le
 
 	SvnStreamWrapper^ sw = AprBaton<SvnStreamWrapper^>::Get((IntPtr)baton);
 
-	array<unsigned char>^ bytes = gcnew array<unsigned char>(*len);
+	array<unsigned char>^ bytes = gcnew array<unsigned char>((int)*len);
 
 	System::Runtime::InteropServices::Marshal::Copy((IntPtr)const_cast<char*>(data), bytes, 0, bytes->Length);
